requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15.0,<3.15.1
django-extensions>=3.2.3
psycopg2==2.9.9
drf-spectacular==0.27.0
Pillow>=8.2.0,<8.3.0
uwsgi>=2.0.19,<2.1
channels==4.1.0
daphne==4.1.2
